Create least_squares_method.py
Created for ELEC 344 Lab 5 Task 5.2 step 2). 

Using the method of Least Squares to fit a linear equation to data points. The equation given is Vdc=IdcReq+ωrKv,eqV​, and it's formulated as a matrix problem b=Ax, where b is a vector containing the measured Vdc​ values, A is a matrix containing the measured Idc and ωr​ values, and x is a vector containing the unknown coefficients Req​ and Kv,eq​.